:bug: Edit to Fix close connection in Reader
 - Also implement new unit test to check close func in JS
 - Refactor to implement js_close in ReaderBase
