+++ src/database/reader_writer/mod.rs
@@ -2,4 +2,4 @@ pub mod read_writer_db;
 pub mod reader_base;
 pub mod reader_db;
 
-pub use reader_base::ReaderBase;
+pub use reader_base::{ReaderBase, SharedReaderBase};
